+++ animation_nodes/nodes/material/object_material_input.py
@@ -1,0 +1,15 @@
+import bpy
+from ... base_types import AnimationNode
+
+class ObjectMaterialInputNode(bpy.types.Node, AnimationNode):
+    bl_idname = "an_ObjectMaterialInputNode"
+    bl_label = "Object Material Input"
+
+    def create(self):
+        self.newInput("Object", "Object", "object", defaultDrawType = "PROPERTY_ONLY")
+        self.newOutput("Material List", "Materials", "materials")
+
+    def execute(self, object):
+        if object is None or not hasattr(object.data, "materials"): return []
+        return [material for material in object.data.materials if material]
+