+++ animation_nodes/ui/node_menu.py
@@ -613,12 +613,13 @@ class MaterialMenu(bpy.types.Menu):
 
     def draw(self, context):
         layout = self.layout
-        insertNode(layout, "an_DataInputNode", "Material", {"assignedType" : repr("Material")})
+        insertNode(layout, "an_DataInputNode", "Material Input", {"assignedType" : repr("Material")})
+        insertNode(layout, "an_ObjectMaterialInputNode", "Object Material Input")
         insertNode(layout, "an_ObjectMaterialOutputNode", "Object Material Output")
-        insertNode(layout, "an_SetPolygonMaterialIndexNode", "Set Polygon Material Index")
-        insertNode(layout, "an_CyclesMaterialOutputNode", "Cycles Material Output")
         insertNode(layout, "an_MaterialOutputNode", "Material Output")
+        insertNode(layout, "an_CyclesMaterialOutputNode", "Cycles Material Output")
         insertNode(layout, "an_GPMaterialOutputNode", "GP Material Output")
+        insertNode(layout, "an_SetPolygonMaterialIndexNode", "Set Polygon Material Index")
 
 class ParticleSystemMenu(bpy.types.Menu):
     bl_idname = "AN_MT_particle_system_menu"